perf(ui-dashboard): SSR homepage + slim per-page GraphQL fetches
Three-part GraphQL efficiency pass driven by network-inspecting the live
dashboard. Measured wins on dev against the same Envio endpoint:

- Homepage: 0 client-side GraphQL requests before first paint. `page.tsx`
  is now an async Server Component that awaits `fetchAllNetworks` and
  hands the payload to the client through `SWRConfig` fallback, so SWR's
  first render reads SSR data rather than firing 14 queries and waiting
  for them. Background revalidation on the hook's `refreshInterval` is
  unchanged. Required extracting `fetchAllNetworks` + all types to
  `src/lib/fetch-all-networks.ts` — swr's react-server build has no
  default export, so a module that both SSR imports and contains the
  hook can't be marked `"use client"`.

- Bridge flows (/bridge-flows): was firing `useAllNetworksData` (7
  queries/chain) just to read `rates` for USD conversion. Replaced with
  new `useOracleRates` hook backed by the slim `ORACLE_RATES` query
  (5 fields × oracleOk pools). Went from 20 requests/mount to 8.

- Revenue (/revenue): was firing `useAllNetworksData` for `fees` +
  `feeTransfers` and discarding the paginated snapshots, LPs, OLS pools,
  trading limits, and breach rollup. Replaced with new `useProtocolFees`
  hook that runs only ORACLE_RATES + PROTOCOL_FEE_TRANSFERS_ALL per
  chain. Went from 14 requests/mount to 4.

- Pool detail FX pools: the cross-pool rate fetch now uses ORACLE_RATES
  instead of ALL_POOLS_WITH_HEALTH — identical rate map, ~85% smaller
  response (5 fields vs 44, filtered to oracleOk).

Inherits PR #202's SWR hardening (no focus/reconnect revalidate, no
hidden-tab refresh, 429-aware retry backoff) on both new hooks.
